requirements.txt
asgiref==3.4.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-on-heroku==1.1.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.1
idna==3.2
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
pyphen==0.11.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.6
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
